Fixed: This plugins did not create the aliases.
The property path dose not exits on the file object in Metalsmith so
I changed it to to use the file name instead. Which actual is not
a filename, it is the url path.

Also I added a .travis.yml so the project can have it's tests run
automatically on the Travic CI server.